+++ src/Stack/Types/Build.hs
@@ -635,7 +635,6 @@ configureOptsNoDir econfig bco deps isLocal package = concat
     , map ("--extra-include-dirs=" ++) (configExtraIncludeDirs config)
     , map ("--extra-lib-dirs=" ++) (configExtraLibDirs config)
     , maybe [] (\customGcc -> ["--with-gcc=" ++ toFilePath customGcc]) (configOverrideGccPath config)
-    , ["--ghcjs" | wc == Ghcjs]
     , ["--exact-configuration"]
     , ["--ghc-option=-fhide-source-paths" | hideSourcePaths cv]
     ]